Move Function classes to a sub dir
