+++ paddle/fluid/jit/layer.cc
@@ -16,8 +16,8 @@
 
 #include "paddle/fluid/framework/variable.h"
 
-#include "paddle/fluid/jit/base_function.h"
 #include "paddle/fluid/jit/compilation_unit.h"
+#include "paddle/fluid/jit/function/base_function.h"
 #include "paddle/fluid/jit/function_schema.h"
 
 namespace paddle {
